+++ NEWS.md
@@ -6,6 +6,7 @@
 * `reactable()` gains a `defaultExpanded` argument to set default expanded rows.
 * New `updateReactable()` function to update the selected rows, expanded rows, or
   current page of a reactable instance in Shiny.
+* New `getReactableState()` function to get the state of a reactable instance in Shiny.
 * In `reactable()`, the `rowClass`, `rowStyle`, and `details` JavaScript functions
   now receive a `rowInfo.selected` property indicating whether the row is selected.